pyproject.toml
[build-system]
requires = ["flit_core >=3.8.0,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "makkulupy"
description = "a lightweight consumer for the makkulu dictionary api"
authors = [{name = "marzka", email = "git@marzka.cafe"}]
readme = "README.md"
dynamic = ["version"]
requires-python = ">= 3.8"
dependencies = [
    "orjson >= 3.8.6, < 3.9",
    "python-dateutil >= 2.8.2, < 2.9",
    "requests >= 2.28, < 2.29"
]
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3"
]

[project.optional-dependencies]
tests = [
    "ward >= 0.67.2b0, < 0.68"
]

[project.urls]
Home = "https://github.com/marzusia/makkulupy"
"Bug Tracker" = "https://github.com/marzusia/makkulupy/issues"